refactor: befit-ai.js 모듈화 진행
