+++ src/befit-ai/modules/constants-befit-ai.js
@@ -1,0 +1,52 @@
+// /src/befit-ai/modules/constants-befit-ai.js
+
+/**
+ * @file BeFit AI 기능에서 사용되는 모든 상수(Constants)를 정의하는 모듈입니다.
+ * @description 이 파일은 애플리케이션 전체에서 재사용되는 고정 값들(예: 스토리지 키, 설정 값, 변환 테이블, 공식 계수 등)을
+ *              한 곳에 모아 관리합니다.
+ */
+
+// [스토리지 키] 로컬 스토리지에 AI 분석 결과를 저장할 때 사용할 키(key)입니다.
+// 이 키를 상수로 관리하면, 여러 파일에서 키 이름을 일관되게 사용할 수 있어 오타로 인한 버그를 방지할 수 있습니다.
+export const STORAGE_KEY = 'befit_ai_result';
+
+// [BMR 모드] 기초대사량(BMR) 입력 방식을 나타내는 값들입니다.
+// 'auto'나 'manual' 같은 문자열을 직접 코드에 쓰는 대신 이 상수를 사용하면,
+// 오타를 방지하고 어떤 모드들이 있는지 명확하게 알 수 있습니다.
+export const BMR_MODE = {
+    AUTO: 'auto',     // 자동 계산 모드
+    MANUAL: 'manual'  // 직접 입력 모드
+};
+
+// [변환 테이블] 사용자가 '상', '중', '하'와 같은 추상적인 수준을 선택했을 때,
+// 이를 실제 수치로 변환하기 위한 기준표입니다.
+export const CONVERSION_TABLES = {
+    // 체지방률(%) 변환표
+    fatRate: {
+        남: { 상: 30, 중상: 25, 중: 20, 중하: 15, 하: 10 },
+        여: { 상: 35, 중상: 30, 중: 27, 중하: 24, 하: 20 },
+        평균: { 상: 32, 중상: 27.5, 중: 23.5, 중하: 19.5, 하: 15 } // 성별 '비공개' 선택 시 사용
+    },
+    // 골격근량(kg)을 계산하기 위한 체중 대비 골격근 비율(%) 변환표
+    skeletalMuscleRate: {
+        남: { 상: 0.45, 중상: 0.40, 중: 0.35, 중하: 0.30, 하: 0.25 }, // 예: 체중 70kg 남성의 '상' 수준은 70 * 0.45 = 31.5kg
+        여: { 상: 0.40, 중상: 0.35, 중: 0.30, 중하: 0.25, 하: 0.20 },
+        평균: { 상: 0.425, 중상: 0.375, 중: 0.325, 중하: 0.275, 하: 0.225 }
+    },
+    // 활동 칼로리(kcal)를 계산하기 위한 체중 대비 활동량 계수 변환표
+    activityRate: {
+        // 예: 체중 70kg 사용자의 '상' 수준 활동량은 70 * 12 = 840kcal
+        상: 12, 중상: 10, 중: 8, 중하: 6, 하: 4
+    }
+};
+
+// [BMR 계산 공식] 기초대사량(BMR)을 자동으로 계산할 때 사용하는 해리스-베네딕트 공식의 변형된 계수들입니다.
+// 각 계수에 이름을 부여하여 공식의 가독성을 높이고, 향후 공식 변경 시 쉽게 대응할 수 있도록 합니다.
+// 공식: (10 * 체중) + (6.25 * 신장) - (5 * 나이) + (성별 상수)
+export const BMR_FORMULA = {
+    WEIGHT_COEFFICIENT: 10,     // 체중 계수
+    HEIGHT_COEFFICIENT: 6.25,   // 신장 계수
+    AGE_COEFFICIENT: 5,         // 나이 계수
+    MALE_CONSTANT: 5,           // 남성 추가 상수
+    FEMALE_CONSTANT: -161       // 여성 추가 상수
+};
